Update ManagerID to use Long and add department update functionality.

+++ src/main/java/edu/icet/demo/dto/records/ManagerID.java
@@ -1,3 +1,3 @@
 package edu.icet.demo.dto.records;
 
-public record ManagerID(int id) {}
+public record ManagerID(Long id) {}
